requirements.txt
Django
django-rest-framework
django-extensions
drupan
django_q
